+++ src/compiler.rs
@@ -1224,6 +1224,7 @@ fn check_for_at_arity(prev_arity: (ArityParam, ArityParam), ast: &VVal,
 
         if let VVal::Lst(l) = vars {
             let llen = l.borrow().len();
+            println!("LI: {:?}", l.borrow());
 
             let var = ast.at(2).unwrap().at(1).unwrap();
             if var.with_s_ref(|var: &str| var == "@") {